Add timer for breaks
Added a timer for breaks.  Uses contenteditable for the break page
heading.  Timer features start with "S" key, pause with "P" key, reset
time with "R" key.  All interactive features are keyboard accessible.
Range slider uses datalist for ticks, which is not supported in Safari.
Not tested in MS Edge.  Base components need to be decoupled.